Refactor: change polymorphic type discriminator for packets
Since these are actually used on the Python side, `$type` isn't a valid attribute name
