+++ afp-ui/Packet/Data/ScopeActionPacketData.cs
@@ -1,4 +1,4 @@
-namespace AFP.Packet;
+namespace AFP.Packet.Data;
 
 public class ScopeActionPacketData : IPacketData
 {
