+++ afp-ui/Packet/Data/ScopeInfoPacketData.cs
@@ -1,4 +1,4 @@
-namespace AFP.Packet;
+namespace AFP.Packet.Data;
 
 public class ScopeInfoPacketData : IPacketData
 {
