+++ afp-ui/Packet/PacketContainer.cs
@@ -1,9 +1,10 @@
 using System.Collections.Generic;
+using AFP.Packet.Data;
 
 namespace AFP.Packet;
 
 public class PacketContainer
 {
-	public required string From { get; set; }
+	public required string Origin { get; set; }
 	public required List<IPacketData> Data { get; set; }
 }
